refactor: split policy facade modules
